+++ src/FirmwareAnalysis/FirmwareAnalysis.Autorest/test/Get-AzFirmwareAnalysisUsageMetric.Tests.ps1
@@ -1,11 +1,11 @@
-if(($null -eq $TestName) -or ($TestName -contains 'Remove-AzFirmwareAnalysisWorkspace'))
+if(($null -eq $TestName) -or ($TestName -contains 'Get-AzFirmwareAnalysisUsageMetric'))
 {
   $loadEnvPath = Join-Path $PSScriptRoot 'loadEnv.ps1'
   if (-Not (Test-Path -Path $loadEnvPath)) {
       $loadEnvPath = Join-Path $PSScriptRoot '..\loadEnv.ps1'
   }
   . ($loadEnvPath)
-  $TestRecordingFile = Join-Path $PSScriptRoot 'Remove-AzFirmwareAnalysisWorkspace.Recording.json'
+  $TestRecordingFile = Join-Path $PSScriptRoot 'Get-AzFirmwareAnalysisUsageMetric.Recording.json'
   $currentPath = $PSScriptRoot
   while(-not $mockingPath) {
       $mockingPath = Get-ChildItem -Path $currentPath -Recurse -Include 'HttpPipelineMocking.ps1' -File
@@ -14,11 +14,12 @@ if(($null -eq $TestName) -or ($TestName -contains 'Remove-AzFirmwareAnalysisWork
   . ($mockingPath | Select-Object -First 1).FullName
 }
 
-Describe 'Remove-AzFirmwareAnalysisWorkspace' {
-    It 'Delete' {
+Describe 'Get-AzFirmwareAnalysisUsageMetric' {
+    It 'Get' {
         { 
-            $config = Remove-AzFirmwareAnalysisWorkspace -ResourceGroupName $env.ResourceGroup -Name 'tobedeleted'
-            $config.Count | Should -eq 0
+            $usage = Get-AzFirmwareAnalysisUsageMetric -ResourceGroupName $env.ResourceGroup -WorkspaceName $env.WorkspaceName -Name 'current'
+            $usage.Count | Should -BeGreaterThan 0
         } | Should -Not -Throw
     }
+
 }