+++ src/FirmwareAnalysis/FirmwareAnalysis/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Upgraded API support to include 2025-08-02
+* Added usage metric support via the Get-AzFirmwareAnalysisUsageMetric cmdlet
 
 ## Version 0.1.4
 * Upgraded nuget package to signed package.